Load 'age' field from the DBI stream, not the PDB stream.
The 'age' field in the PDB stream changes each time the PDB gets
modified by e.g. PDBSTR (for source indexing). However, the signature
needs to use the 'age' when the PDB file was linked. This 'age' is
found in the DBI stream (stream 3). With this change, the signature
produced is the same as 'symstore.exe''s output.