add delete transaction operation
Adds support for deleting a specified transaction
from a symstore.

On delete, the transaction will be marked as deleted
in the meta files.

The symbols files will be garbage collected. That is,
any files that are not refered by existing transaction,
they will be deleted.

Adds command line '--delete' option, to invoke delete
operation from CLI.